+++ docs/pages/web/indicator.tsx
@@ -1,0 +1,53 @@
+import docGen, { DocGen } from '../../docs-components/docgen';
+import GeneratedPropTable from '../../docs-components/GeneratedPropTable';
+import LocalizationSection from '../../docs-components/LocalizationSection';
+import MainSection from '../../docs-components/MainSection';
+import Page from '../../docs-components/Page';
+import PageHeader from '../../docs-components/PageHeader';
+import SandpackExample from '../../docs-components/SandpackExample';
+import counter from '../../examples/indicator/counter';
+import main from '../../examples/indicator/main';
+import notification from '../../examples/indicator/notification';
+
+export default function DocsPage({ generatedDocGen }: { generatedDocGen: DocGen }) {
+  return (
+    <Page title={generatedDocGen?.displayName}>
+      <PageHeader description={generatedDocGen?.description} name={generatedDocGen?.displayName}>
+        <SandpackExample
+          code={main}
+          hideEditor
+          name={`Main ${generatedDocGen?.displayName} example`}
+        />
+      </PageHeader>
+
+      <GeneratedPropTable generatedDocGen={generatedDocGen} />
+
+      <LocalizationSection name={generatedDocGen?.displayName} noDefaultLabelProvider />
+
+      <MainSection name="Variants">
+        <MainSection.Subsection description="Notification" title="Notification">
+          <MainSection.Card
+            cardSize="lg"
+            sandpackExample={<SandpackExample code={notification} name="Notification example" />}
+          />
+        </MainSection.Subsection>
+        <MainSection.Subsection description="Counter" title="Counter">
+          <MainSection.Card
+            cardSize="lg"
+            sandpackExample={<SandpackExample code={counter} name="Counter example" />}
+          />
+        </MainSection.Subsection>
+      </MainSection>
+    </Page>
+  );
+}
+
+export async function getServerSideProps(): Promise<{
+  props: {
+    generatedDocGen: DocGen;
+  };
+}> {
+  return {
+    props: { generatedDocGen: await docGen('Indicator') },
+  };
+}