+++ packages/gestalt/src/Badge.tsx
@@ -13,8 +13,6 @@ import useInExperiment from './useInExperiment';
 import useInteractiveStates from './utils/useInteractiveStates';
 import { Indexable } from './zIndex';
 
-type Position = 'middle' | 'top';
-
 type TooltipProps = {
   accessibilityLabel?: string;
   idealDirection?: 'up' | 'right' | 'down' | 'left';
@@ -50,7 +48,7 @@ type Props = {
   /**
    * Badge position relative to its parent element. See the [positioning](https://gestalt.pinterest.systems/web/badge#Positioning) variant to learn more.
    */
-  position?: Position;
+  position?: 'middle' | 'top';
   /**
    * Text displayed inside of the Badge. Sentence case is preferred.
    */