+++ packages/gestalt/src/contexts/DefaultLabelProvider.tsx
@@ -99,6 +99,9 @@ export type DefaultLabelContextType = {
   Spinner: {
     accessibilityLabel: string;
   };
+  Tabs: {
+    accessibilityNotificationLabel: string;
+  };
   TableOfContents: {
     accessibilityLabel: string;
   };
@@ -229,6 +232,9 @@ export const fallbackLabels: DefaultLabelContextType = {
   TableOfContents: {
     accessibilityLabel: 'Table of contents',
   },
+  Tabs: {
+    accessibilityNotificationLabel: 'This tab is displaying a notification indicator',
+  },
   Tag: {
     accessibilityErrorIconLabel: 'Error',
     accessibilityRemoveIconLabel: 'Remove tag',
@@ -244,6 +250,7 @@ export const fallbackLabels: DefaultLabelContextType = {
   HelpButton: {
     tooltipMessage: 'Click to learn more',
   },
+
   Toast: {
     accessibilityDismissButtonLabel: 'Dismiss message',
     accessibilityIconSuccessLabel: 'Success message',